✨ feat(project-dashboard-card): refactor to consolidate ranked list display logic into project-dashboard-card component, removing the deprecated ranked list component; enhance layout options and integrate dynamic bar color and width calculations for improved visualization
