+++ research-indicators/src/app/pages/platform/pages/project-detail/components/project-dashboard-card/project-dashboard-card.component.html
@@ -1,3 +1,6 @@
+@if (variant() === 'list') {
+<ng-container *ngTemplateOutlet="rankedList"></ng-container>
+} @else {
 <section class="box-border flex h-full flex-col rounded-xl border border-[#e8ebed] bg-[#f4f7f9] p-5 shadow-xs"
   [class.min-h-0]="compact()" [class.min-h-[280px]]="!compact()">
   <header class="mb-4 flex items-start justify-between gap-3">
@@ -27,6 +30,10 @@
       <i class="pi pi-chart-bar text-[28px] text-[#1689CA]" aria-hidden="true"></i>
       <p class="m-0 max-w-64 font-['Barlow'] text-[13px] leading-snug text-[#777C83]">{{ emptyMessage() }}</p>
     </div>
+    } @else if (items().length) {
+    <div class="flex min-h-0 min-w-0 flex-1 flex-col overflow-hidden">
+      <ng-container *ngTemplateOutlet="rankedList"></ng-container>
+    </div>
     } @else {
     <div class="flex min-h-0 min-w-0 flex-1 flex-col overflow-hidden">
       <ng-content></ng-content>
@@ -34,3 +41,113 @@
     }
   </div>
 </section>
+}
+
+<ng-template #rankedList>
+  @if (layout() === 'columns') {
+  <ul
+    class="m-0 grid w-full max-w-full list-none box-border items-end gap-1.5 overflow-hidden rounded-[10px] bg-[#f4f7f9] px-2 pb-2.5 pt-3 p-0"
+    [style.min-height.px]="largeColumns() ? 200 : 140" [class.justify-center]="items().length <= 3"
+    [style.grid-template-columns]="
+      items().length <= 3
+        ? 'repeat(' + items().length + ', minmax(0, ' + (largeColumns() ? '72px' : '52px') + '))'
+        : 'repeat(' + items().length + ', minmax(0, 1fr))'
+    ">
+    @for (item of items(); track item.id; let index = $index) {
+    <li class="flex w-full min-w-0 flex-col items-center gap-1.5">
+      <span class="font-['Barlow'] text-xs font-bold leading-none tabular-nums text-[#112F5C]">{{ item.count }}</span>
+      <div class="relative mx-auto w-full overflow-hidden" [style.height.rem]="largeColumns() ? 9 : 6"
+        [style.max-width.rem]="4.2" aria-hidden="true">
+        <div class="absolute inset-x-0 bottom-0 min-h-1 w-full rounded-t-lg transition-[height] duration-300 ease-in-out"
+          [style.background-color]="barColor(index)" [style.height.%]="fillPercent(item.count)"></div>
+      </div>
+      <span
+        class="w-full cursor-default overflow-hidden break-words text-center font-['Barlow'] leading-tight line-clamp-2"
+        [class.text-xs]="largeColumns()" [class.text-[10px]]="!largeColumns()"
+        [style.color]="largeColumns() ? '#777C83' : '#4c5158'" [pTooltip]="item.label" tooltipPosition="bottom">{{
+        item.label }}</span>
+    </li>
+    }
+  </ul>
+  } @else if (layout() === 'rows-partners') {
+  <ul class="m-0 flex list-none flex-col gap-3 justify-center pt-1">
+    @for (item of items(); track item.id; let index = $index) {
+    <li class="grid grid-cols-[minmax(0,1fr)_minmax(72px,1.5fr)] items-center gap-2.5">
+      <span
+        class="min-w-0 cursor-default overflow-hidden font-['Barlow'] text-[12.5px] leading-tight text-[#4c5158] font-semibold line-clamp-2"
+        [appTruncatedTooltip]="item.label" tooltipPosition="top">{{ item.label }}</span>
+      <div class="flex min-w-0 w-full items-center overflow-visible">
+        <div class="h-6 shrink-0 rounded-md transition-[width] duration-300 ease-in-out"
+          [style.background-color]="barColor(index)" [style.width.%]="partnerBarWidthPercent(item.count)"
+          [style.min-width.px]="item.count > 0 ? 8 : 0" aria-hidden="true"></div>
+        <span class="ml-1.5 shrink-0 font-['Barlow'] text-xs font-bold leading-none tabular-nums text-[#112F5C]">{{
+          item.count }}</span>
+      </div>
+    </li>
+    }
+  </ul>
+  } @else if (layout() === 'rows-stacked-lever') {
+  <div class="flex w-full flex-1 flex-col min-h-0" [class.justify-center]="items().length >= 2"
+    [class.pt-4]="items().length < 2" [class.justify-start]="items().length < 2">
+    <ul class="m-0 flex w-full list-none flex-col gap-3.5 p-0">
+      @for (item of items(); track item.id; let index = $index) {
+      <li class="flex flex-col gap-2.5 rounded-md border border-[#e8ebed] bg-white/80 px-5 py-2.5">
+        <div class="flex min-w-0 items-center gap-2.5">
+          @if (item.iconUrl) {
+          <img [src]="item.iconUrl" alt="" class="h-7 w-7 shrink-0 rounded-full object-cover ring-1 ring-black/[0.06]"
+            aria-hidden="true" />
+          }
+          <span
+            class="min-w-0 cursor-default overflow-hidden break-words font-['Barlow'] text-[11px] font-semibold uppercase leading-snug tracking-wide text-[#4c5158] line-clamp-2"
+            [appTruncatedTooltip]="item.label" tooltipPosition="top">{{ item.label }}</span>
+        </div>
+        <div class="flex items-center gap-2.5">
+          <div class="h-2.5 min-w-0 flex-1 overflow-hidden rounded-full bg-[#e8ebed]" aria-hidden="true">
+            <div class="h-full min-w-1 rounded-[inherit] transition-[width] duration-300 ease-in-out"
+              [style.background-color]="barColor(index)" [style.width.%]="fillPercent(item.count)"></div>
+          </div>
+          <span class="min-w-7 text-right font-['Barlow'] text-[13px] font-bold tabular-nums text-[#112F5C]">{{ item.count
+            }}</span>
+        </div>
+      </li>
+      }
+    </ul>
+  </div>
+  } @else if (layout() === 'rows-stacked') {
+  <div class="flex w-full flex-1 flex-col min-h-0" [class.justify-center]="items().length >= 2"
+    [class.justify-start]="items().length < 2">
+    <ul class="m-0 flex w-full list-none flex-col gap-3.5 p-0">
+      @for (item of items(); track item.id; let index = $index) {
+      <li class="flex flex-col gap-2.5 rounded-md border border-[#e8ebed] bg-white/80 px-5 py-2.5">
+        <span
+          class="cursor-default overflow-hidden break-words font-['Barlow'] text-sm font-semibold leading-snug text-[#4c5158] line-clamp-2"
+          [appTruncatedTooltip]="item.label" tooltipPosition="top">{{ item.label }}</span>
+        <div class="flex items-center gap-2.5">
+          <div class="h-2.5 min-w-0 flex-1 overflow-hidden rounded-full bg-[#e8ebed]" aria-hidden="true">
+            <div class="h-full min-w-1 rounded-[inherit] transition-[width] duration-300 ease-in-out"
+              [style.background-color]="barColor(index)" [style.width.%]="fillPercent(item.count)"></div>
+          </div>
+          <span class="min-w-7 text-right font-['Barlow'] text-[13px] font-bold tabular-nums text-[#112F5C]">{{ item.count
+            }}</span>
+        </div>
+      </li>
+      }
+    </ul>
+  </div>
+  } @else {
+  <ul class="m-0 flex list-none flex-col gap-2.5 p-0">
+    @for (item of items(); track item.id; let index = $index) {
+    <li class="grid grid-cols-[minmax(0,1fr)_minmax(64px,1.6fr)_auto] items-center gap-2.5">
+      <span class="cursor-default truncate font-['Barlow'] text-xs text-[#4c5158]" [pTooltip]="item.label"
+        tooltipPosition="top">{{ item.label }}</span>
+      <div class="h-2.5 overflow-hidden rounded-full bg-[#e8ebed]" aria-hidden="true">
+        <div class="h-full min-w-1 rounded-[inherit] transition-[width] duration-300 ease-in-out"
+          [style.background-color]="barColor(index)" [style.width.%]="fillPercent(item.count)"></div>
+      </div>
+      <span class="min-w-7 text-right font-['Barlow'] text-[13px] font-bold tabular-nums text-[#112F5C]">{{ item.count
+        }}</span>
+    </li>
+    }
+  </ul>
+  }
+</ng-template>
